+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - feature/1199516440696130 - Add implode method to string collection / Add removeByValue method (with optional strict comparison)
 - feature/1199718460902669 - MultipleSelectionInterface (and trait) / Improve getSelectedItems method
 - feature/1199601516820605 - Add hasSelection to MultipleSelection and SingleSelection (interfaces and traits)
+- feature/1199692948932206 - Add ClassStringCollection (Collection of classes as a string with existence validation)
 
 ### Changed